Assignment 2 Code
Here, including the digest library, I compute the md5 checksum of the
invertible matrix passed to the cacheSolve() function.
cacheSolve()
If the md5 doesn’t exist in the present environment, then the original
matrix hasn’t been solved, and makeCacheMatrix is called. If it does
exist, the solved matrix is returned.
makeCacheMatrix()
Create a variable to store the solution using md5 of the original
matrix. Solve the matrix (assumes it is invertible), store the solution
to the variable, and display the solution.